♻️ refactor: separate form initial state for better responsibility
- Moved initialFormState out of formReducer for clearer separation
- Initial state now resides in a separate file to define data structure
- Improved maintainability and testability by separating logic and structure
